chore: update Icon types
DSY-1965
